+++ ckeditor5_plugins/invisible/src/invisibleediting.js
@@ -1,7 +1,7 @@
 import { Plugin } from 'ckeditor5/src/core';
 import InvisibleCommand from './insertinvisiblecommand';
-import { Widget } from 'ckeditor5/src/widget';
-
+import RemoveInvisibleCommand from './removeinvisiblecommand';
+import { Widget, toWidget } from 'ckeditor5/src/widget';
 
 export default class InvisibleEditing extends Plugin {
 	static get requires() {
@@ -15,47 +15,55 @@ export default class InvisibleEditing extends Plugin {
 			'addInvisible',
 			new InvisibleCommand(this.editor),
 		  );
+		this.editor.commands.add(
+			'removeInvisible',
+			new RemoveInvisibleCommand(this.editor)
+		)
 	}
 	_defineSchema() {
 		const schema = this.editor.model.schema;
-    	// Extend the text node's schema to accept the tooltip attribute.
+	
 		schema.register( 'ucb-invisible', {
-			allowWhere: '$block',
-			allowChildren: '$text'
-		} );
+			isObject: true,
+			allowWhere: '$text',
+			allowContentOf: '$block', // It can contain content of a block (allows text implicitly)	
+			isInline: true,
+		});
 	}
+	
 	_defineConverters() {
-		const {conversion} = this.editor;
-
-		// Conversion from a view element to a model attribute
-		conversion.for( 'upcast' ).elementToElement( {
-			model: 'ucb-invisible',
+		const { conversion } = this.editor;
+	
+		conversion.for('upcast').elementToElement({
+			model: (viewElement, { writer }) => {
+				return writer.createElement('ucb-invisible');
+			},
 			view: {
 				name: 'span',
-				classes:  'sr-only'
+				classes: 'sr-only'
 			},
-		} );
-
-        // Conversion from a model attribute to a view element
-		conversion.for( 'dataDowncast' ).elementToElement( {
+		});
+	
+		conversion.for('dataDowncast').elementToElement({
 			model: 'ucb-invisible',
-
-            // Callback function provides access to the model attribute value
-			// and the DowncastWriter
 			view: {
 				name: 'span',
 				classes: 'sr-only',
-			}
-		} );
-
+			},
+		});
+	
 		conversion.for('editingDowncast').elementToElement({
 			model: 'ucb-invisible',
-			view: {
-				name: 'span',
-				classes: 'ucb-invisible'
-			}
-		})
-
-
+			view: (modelItem, { writer }) => {
+				const widgetElement = createInvisibleWidget(writer);
+				return widgetElement;
+			},
+		});
 	}
+	
+}
+
+function createInvisibleWidget(writer) {
+	const span = writer.createContainerElement('span', { class: 'ucb-invisible' });
+	return toWidget(span, writer, { label: 'invisible widget',hasSelectionHandle: true },);
 }